Table 4 - 46 arterial blood gas measurements.

A dataset containing 46 arterial blood gas measurements from four patients. Presented as table 4 in (Douglas et al. 1988) from the work of (Jones et al. 1971)

Usage

douglas_table_4

Format

A data frame with 48 rows and 8 variables:

subject

subject, of four participants

po2_torr

partial pressure of O2 in torr(mmHg), contains some non numeric entries!

ph

ph, hydrogen ion concentration

haemoglobin_g_dl

haemoglobin, in grams per decilitre

so2_fraction

haemoglobin oxygen saturation as a fraction, ie between 0 and 1

blood_co2_content_ml_dl

CO2 content of blood in mls per decilitre

pco2_torr

partial pressure of CO2 in torr(mmHg)

plasma_co2_content_ml_dl

CO2 content of plasma in mls per decilitre

References

Douglas AR, Jones NL, Reed JW (1988). “Calculation of whole blood CO2 content.” J. Appl. Physiol., 65(1), 473–477. Jones NL, Levine GB, Robertson DG, Epstein SW (1971). “The effect of added dead space on the pulmonary response to exercise.” Respiration, 28(5), 389–398.